Movendo o botao de criar novo conteudo para a toolbar
